Added DFTB+ geom. opt. to ase_phonon_calc
ase_phonon_calc will now geometry optimize the input cell using DFTB+,
so that phonon modes can be accurately calculated using DFTB+ which
would not be the case if just CASTEP optimization was used. The
get_major_emodes function can now orthogonolize and normalize the major
modes before returning them which is also important for DFTB+ phonons.
Once calculated, the DFTB+ phonon modes will be used to displace the
atoms of the /original/ structure that was input.